posts/comments sequelize
First draft for posts n comments sequelize, as well as some additional files for future use
