+++ models/User.js
@@ -1,0 +1,42 @@
+const {Model, DataType, DataTypes}= require("sequelize");
+const sequelize = require("../config/connection")
+
+class User extends Model{};
+
+User.init(
+    {
+        id:{
+            type :DataTypes.INTEGER,
+            allowNull:false,
+            primaryKey: true,
+            autoIncrement:true
+        },
+        name:{
+            type: DataTypes.STRING,
+            allowNull:false,
+        },
+        email:{
+            type:DataTypes.STRING,
+            allowNull: false,
+            unique:true,
+            validate:{
+                isEmail:true,
+            }
+        },
+        password:{
+            type:DataTypes.STRING,
+            allowNull:false,
+            validate:{
+                len:[6],
+            }
+        }
+
+    },
+    {
+        sequelize,
+        freezeTableName :true,
+        underscored:true,
+        modelName:'user'
+    }
+);
+module.exports= User;
