+++ views/Layouts/main.handlebars
@@ -1,0 +1,19 @@
+<html lang='en'>
+<head>
+
+</head>
+
+<body>
+    <div class='flex-column justify-space-around'>
+    <header class='display-flex justify-space-between align-center p-2'>
+        <h1>
+            PAC-Overflow
+        </h1>
+    </header>
+    <main>
+        {{body}}
+    </main>
+    </div>
+</body>
+
+</html>
